[Chore #163239568] Refactor meetup questions specific API feature to new file
