+++ server/middlewares/index.js
@@ -1,0 +1,8 @@
+import Auth from './auth';
+import Misc from './misc';
+import schemaValidator from './schema/schemaValidator';
+import Upload from './uploads';
+
+export {
+  Auth, Misc, Upload, schemaValidator
+};
